+++ src/app/api/plugins/context.rs
@@ -63,6 +63,7 @@ impl App {
                         })
                 }),
             EventData::WorkspaceRenamed { workspace_id, .. }
+            | EventData::WorkspaceMoved { workspace_id, .. }
             | EventData::WorkspaceFocused { workspace_id } => self
                 .plugin_context_for_workspace_id(workspace_id, correlation_id)
                 .unwrap_or_else(|| {
@@ -99,6 +100,11 @@ impl App {
                 context
             }
             EventData::TabRenamed {
+                tab_id,
+                workspace_id,
+                ..
+            }
+            | EventData::TabMoved {
                 tab_id,
                 workspace_id,
                 ..